--- apps/expo/index.ts
@@ -1,3 +1,0 @@
-// This is the entry point for the Expo app.
-// It is responsible for setting up the router, root component and loading pages.
-import "expo-router/entry";
