+++ apps/expo/index.tsx
@@ -1,0 +1,13 @@
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+import { registerRootComponent } from "expo";
+import { ExpoRoot } from "expo-router";
+
+// Must be exported or Fast Refresh won't update the context
+export function App() {
+  // @ts-expect-error - require.context is not defined?
+  const ctx = require.context("./src/app");
+  return <ExpoRoot context={ctx} />;
+}
+
+registerRootComponent(App);
