+++ apps/expo/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@acme/expo",
   "version": "0.1.0",
-  "main": "index.ts",
+  "main": "index.tsx",
   "scripts": {
     "clean": "git clean -xdf .expo .turbo node_modules",
     "dev": "expo start --ios",
