+++ apps/expo/src/app/_layout.tsx
@@ -3,7 +3,7 @@ import { SafeAreaProvider } from "react-native-safe-area-context";
 import { Stack } from "expo-router";
 import { StatusBar } from "expo-status-bar";
 
-import { TRPCProvider } from "../src/utils/api";
+import { TRPCProvider } from "../utils/api";
 
 // This is the main layout of the app
 // It wraps your pages with the providers they need
