+++ apps/expo/src/app/post/[id].tsx
@@ -1,7 +1,7 @@
 import { SafeAreaView, Text, View } from "react-native";
 import { SplashScreen, Stack, useSearchParams } from "expo-router";
 
-import { api } from "../../src/utils/api";
+import { api } from "../../utils/api";
 
 const Post: React.FC = () => {
   const { id } = useSearchParams();
