+++ apps/expo/tsconfig.json
@@ -4,6 +4,6 @@
     "jsx": "react-native",
     "types": ["nativewind/types"]
   },
-  "include": ["app", "expo-plugins", "src", "app.config.ts", "index.ts"],
+  "include": ["expo-plugins", "src", "app.config.ts", "index.tsx"],
   "exclude": ["node_modules"]
 }
