+++ turbo.json
@@ -36,6 +36,7 @@
     "DATABASE_URL",
     "DISCORD_CLIENT_ID",
     "DISCORD_CLIENT_SECRET",
+    "EXPO_ROUTER_APP_ROOT",
     "NEXTAUTH_SECRET",
     "NEXTAUTH_URL",
     "NODE_ENV",
